% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-dir.R
\name{detect_mba_format}
\alias{detect_mba_format}
\title{Try to detect the format of a file}
\usage{
detect_mba_format(filepath, format = NULL)
}
\arguments{
\item{filepath}{(\code{character(1)}) The path to the file.}

\item{format}{(\code{character(1)}, optional) If not \code{NULL},
the function will check if the provided format is valid and return it.
If \code{NULL}, the function will attempt to detect the format based on the filename.
The default is \code{NULL}.}
}
\description{
Try to detect the format of a file
}
