% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-bioplex.R
\name{extract_header_information}
\alias{extract_header_information}
\title{Extract header information from the first column of a BIOPLEX file}
\usage{
extract_header_information(first_col)
}
\arguments{
\item{first_col}{The first column of the BIOPLEX excel sheet}
}
\value{
A named list containing the header key-value pairs
}
\description{
Extract header information from the first column of a BIOPLEX file
}
\keyword{internal}
