% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_xlab}
\alias{format_xlab}
\title{Format X-axis label}
\usage{
format_xlab(prefix, prefix_short, scale_y)
}
\arguments{
\item{prefix}{A string indicating the prefix for the x-axis label.}

\item{prefix_short}{A shorter version of the prefix for the x-axis label.}

\item{scale_y}{A string indicating the scale of the x-axis (e.g., "log10", "linear").}
}
\value{
A formatted x-axis label.
}
\description{
Returns a generic x-axis label for the plot based on the prefix and scale.
}
\keyword{internal}
