% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_ylab}
\alias{format_ylab}
\title{Format Y-axis label}
\usage{
format_ylab(data_type, scale_y)
}
\arguments{
\item{data_type}{A string indicating the type of data (e.g., "Median", "Mean").}

\item{scale_y}{A string indicating the scale of the y-axis (e.g., "log10", "linear").}
}
\value{
A formatted y-axis label.
}
\description{
Returns a generic y-axis label for the plot based on the data type and scale.
}
\keyword{internal}
