% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-bioplex.R
\name{get_header_end_idx}
\alias{get_header_end_idx}
\title{Get the index of the end of the header in the first column}
\usage{
get_header_end_idx(first_col)
}
\arguments{
\item{first_col}{The first column of the BIOPLEX excel sheet}
}
\value{
The index of the last row of the header
}
\description{
Get the index of the end of the header in the first column
}
\keyword{internal}
