% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-bioplex.R
\name{read_bioplex_format}
\alias{read_bioplex_format}
\title{Read the BIOPLEX format data}
\usage{
read_bioplex_format(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to the BIOPLEX file (excel file)}

\item{verbose}{Print additional information. Default is \code{TRUE}}
}
\description{
Read the BIOPLEX format data
}
