% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_csv}
\alias{save_csv}
\title{Wrapper for write.table to save CSV files}
\usage{
save_csv(df, filepath, row_names_col = "")
}
\arguments{
\item{df}{(\code{data.frame}) A dataframe to be saved as a CSV file.}

\item{filepath}{(\code{character(1)}) The path where the CSV file will be saved.}

\item{row_names_col}{(\code{character(1)}) The name of the column to store row names. If empty, row names are not saved.}
}
\description{
This wrapper handles locale difference when saving CSV files.
}
\keyword{internal}
