% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{try_cast_as_numeric}
\alias{try_cast_as_numeric}
\title{Try cast dataframe columns as numeric}
\usage{
try_cast_as_numeric(dataframe)
}
\arguments{
\item{dataframe}{(\code{data.frame}) A dataframe whose columns are to be converted to numeric.}
}
\value{
A dataframe with columns converted to numeric where possible.
}
\description{
This function attempts to convert each column of a dataframe to numeric.
Additionally, it replaces commas with dots to handle decimal separators.

If at any point of the conversion a Nan value is detected,
where it was not present in the original column, then
the original column is retained.
}
\keyword{internal}
